PR리뷰 적용 - Cards 클래스를 생성해서 관련 로직 넘김 https://github.com/next-step/kotlin-blackjack/pull/549#discussion_r1255673244
