+++ src/main/kotlin/blackjack/domain/Player.kt
@@ -1,30 +1,9 @@
 package blackjack.domain
 
 class Player(val name: String) {
-    val cards: MutableList<Card> = mutableListOf()
+    val cards = Cards()
 
     fun addCard(card: Card) {
         cards.add(card)
-    }
-
-    fun canDrawMoreCard(): Boolean {
-        return calculateScore() < BLACKJACK
-    }
-
-    fun calculateScore(): Int {
-        val hasAce = cards.any { it.denomination == Denomination.ACE }
-
-        val totalScore = cards.sumOf { it.denomination.score }
-        if (hasAce && totalScore <= BLACKJACK - ACE_SCORE_DIFF) {
-            return totalScore + ACE_SCORE_DIFF
-        }
-        return totalScore
-    }
-
-    companion object {
-        const val BLACKJACK = 21
-        private const val ACE_SCORE_ONE = 1
-        private const val ACE_SCORE_ELEVEN = 11
-        const val ACE_SCORE_DIFF = ACE_SCORE_ELEVEN - ACE_SCORE_ONE
     }
 }
